consistent use of size_t for length parameters

+++ include/sw842.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 
-int sw842_compress(const uint8_t *in, unsigned int ilen,
-		   uint8_t *out, unsigned int *olen);
+int sw842_compress(const uint8_t *in, size_t ilen,
+		   uint8_t *out, size_t *olen);
 
-int sw842_decompress(const uint8_t *in, unsigned int ilen,
-		     uint8_t *out, unsigned int *olen);
+int sw842_decompress(const uint8_t *in, size_t ilen,
+		     uint8_t *out, size_t *olen);
 
 #endif
