+++ src/serial/842_decompress.c
@@ -273,8 +273,8 @@ static int do_op(struct sw842_param_decomp *p, uint8_t o)
  * will contain the number of output bytes written on success, or
  * 0 on error.
  */
-int sw842_decompress(const uint8_t *in, unsigned int ilen,
-		     uint8_t *out, unsigned int *olen)
+int sw842_decompress(const uint8_t *in, size_t ilen,
+		     uint8_t *out, size_t *olen)
 {
 	struct sw842_param_decomp p;
 	int ret;
